fix(auth-files): support antigravity usage trends and render system calls as badge
